Added AtomicRobotUI.kt file for common reusable Composables, consolidated bar Composables to AppBars.kt, removed some unused utils and Composables, and organized a bit.
