--- app/src/main/java/com/atomicrobot/carbon/navigation/TopBar.kt
@@ -1,29 +1,0 @@
-package com.atomicrobot.carbon.navigation
-
-import androidx.compose.material.Icon
-import androidx.compose.material.IconButton
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Text
-import androidx.compose.material.TopAppBar
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.vector.ImageVector
-
-@Composable
-fun TopBar(title: String = "", buttonIcon: ImageVector, onButtonClicked: () -> Unit) {
-    TopAppBar(
-        title = {
-            Text(
-                text = title
-            )
-        },
-        navigationIcon = {
-            IconButton(onClick = { onButtonClicked() }) {
-                Icon(buttonIcon, contentDescription = "")
-            }
-        },
-        backgroundColor = MaterialTheme.colors.primaryVariant
-    )
-}
-
-@Composable
-fun TopBar(title: String = "") = TopAppBar(title = { Text(text = title) })
